swapped CreateConsultation and ModelPrediction because of switched naming
