+++ frontend/src/Pages/Doctor/CreateConsultationPage.jsx
@@ -3,15 +3,14 @@ import React from "react";
 import ActorNavbar from "../../Components/ActorNavbar.jsx";
 import {useParams} from "react-router-dom";
 import CreateConsultation from "../../Components/Doctor/Consultation/Create/CreateConsultation.jsx";
-
 function CreateConsultationPage(){
     const {patientId,consultationId} = useParams();
-
-    return(
+    return (
         <div>
             <ActorNavbar/>
             <CreateConsultation patientId={patientId} consultationId={consultationId}/>
         </div>
     )
 }
+
 export default CreateConsultationPage;
